Fixed memory bugs and formatted AR code in Results view controller. READ FULL COMMIT FOR INFO ON NEW BUG.
Pushing all the commits now because there's too many changes locally for me to feel comfortable. The new bug occurs when switching to the AR view and has something to do with the view type of the PropertyArViewController needing to be an Image Picker class. Will investigate.

+++ Classes/PropertyResultsViewController.h
@@ -43,7 +43,7 @@
 		// Misc
         UIAlertView *alertView_;
 		UISegmentedControl *segmentedControl_;
-		int previousSelectedSegment_;
+		NSInteger previousSelectedSegment_;
 		UIImagePickerController *camera_;
 }
